pyproject.toml
[project]
name = "mcp-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-anthropic>=0.3.21",
    "langchain-mcp-adapters>=0.1.10",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
]
